refactor : Member 인터페이스 선언, AnonymousUser User 구현
