+++ src/main/java/wooteco/auth/domain/AnonymousUser.java
@@ -1,0 +1,9 @@
+package wooteco.auth.domain;
+
+public class AnonymousUser implements User{
+
+    @Override
+    public int calculateFee(int distance, int extraFare) {
+        return defaultCalculator.calculateFare(distance, extraFare);
+    }
+}
