+++ src/main/java/wooteco/subway/domain/fareCalculator/AgeDiscountFareCalculator.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service;
+package wooteco.subway.domain.fareCalculator;
 
 import java.util.Arrays;
 import java.util.function.Predicate;
