+++ src/main/java/wooteco/subway/service/PathService.java
@@ -3,11 +3,10 @@ package wooteco.subway.service;
 import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.auth.dao.MemberDao;
 import wooteco.auth.domain.LoginMember;
-import wooteco.auth.domain.Member;
+import wooteco.auth.domain.User;
+import wooteco.auth.service.MemberService;
 import wooteco.common.exception.badrequest.InvalidPathException;
-import wooteco.common.exception.notfound.MemberNotFoundException;
 import wooteco.subway.domain.Line;
 import wooteco.subway.domain.Station;
 import wooteco.subway.domain.SubwayPath;
@@ -21,16 +20,15 @@ public class PathService {
     private final LineService lineService;
     private final StationService stationService;
     private final PathFinder pathFinder;
-    private final FareCalculator fareCalculator;
-    private final MemberDao memberDao;
+    private final MemberService memberService;
 
     public PathService(LineService lineService, StationService stationService,
-        PathFinder pathFinder, FareCalculator fareCalculator, MemberDao memberDao) {
+        PathFinder pathFinder,
+        MemberService memberService) {
         this.lineService = lineService;
         this.stationService = stationService;
         this.pathFinder = pathFinder;
-        this.fareCalculator = fareCalculator;
-        this.memberDao = memberDao;
+        this.memberService = memberService;
     }
 
     public PathResponse findPath(Long source, Long target, LoginMember loginMember) {
@@ -39,24 +37,16 @@ public class PathService {
             Station sourceStation = stationService.findStationById(source);
             Station targetStation = stationService.findStationById(target);
             SubwayPath subwayPath = pathFinder.findPath(lines, sourceStation, targetStation);
-            FareCalculator fareCalculator = selectFareCalculatorByUser(loginMember);
 
-            final int fare = fareCalculator.calculateFare(
+            final User user = memberService.selectMemberOrAnonymous(loginMember);
+            final int fare = user.calculateFee(
                 subwayPath.calculateDistance(),
                 subwayPath.calculateMaxExtraFare()
             );
+
             return PathResponseAssembler.assemble(subwayPath, fare);
         } catch (Exception e) {
             throw new InvalidPathException();
         }
-    }
-
-    private FareCalculator selectFareCalculatorByUser(LoginMember loginMember) {
-        if (loginMember.isAnonymous()) {
-            return fareCalculator;
-        }
-        final Member member = memberDao.findById(loginMember.getId())
-            .orElseThrow(MemberNotFoundException::new);
-        return new AgeDiscountFareCalculator(fareCalculator, member.getAge());
     }
 }
